--allow-unauthenticated for rethinkdb install
The rethinkdb GPG key seems to have expired. The project has been abandoned so that makes sense. 

Let's try to install it using the --allow-unauthenticated flag so that it will succeed anyway.
Please note that this is a temporary solution justified by the fact that we're quite close to the django 
rewrite release.