Make API return infraction list with newest first
The most recent infraction a user has is usually the most relevant to
us. This PR aims to make our life easier by sorting the infractions by
their `inserted_at` date, with the most recent insertion date first.

Note: The `id` is not entirely in chronological order, because we did
not import historical infractions in chronological order. That's why
the `inserted_at` field is specified instead of the `id` field.