Migrate hidden warnings to notes
https://github.com/python-discord/site/issues/260

We have added a new "note" type to the Infraction model, meaning that
we can now decouple notes and warnings completely. However, notes
made prior to this change are still registered as "hidden warnings".

This commit adds a data migration that sets the type field of "hidden
warning" infractions to "note".

This commit implements the first part of #260