Migrate nominations to new Nomination model
Before the migration to Django, we stored meta data on a nomination,
such as the `reason` and `end_reason`, in the infraction table using
"note" infractions, using a special prefix for the `reason`.

We have since decided to move nominations out of the infraction
context by creating a special `Nomination` model. However, given the
complexity of the data migration, we did not yet migrate the old
nomination data to this new model. This commit migrates that data by performing a data migration.

The data migration works as follows:

- Query all nomination data in chronological order;

- Replay all nominations and add them to the `Nomination` model;

- Delete the now obsolete `Infraction` entry.

In addition, this commit also adds a useful string representation for
`Nomination` objects.