Set SECRET_KEY to constant in debug mode
The SECRET_KEY in debug mode was auto-generated each time the config
file was loaded. This had the unwanted side-effect of requiring
testers to login again on the site as the SECRET_KEY had changed
during the reload process.

This commit resolves that by turning it into a constant. Since having
a constant secret results in the linting error `S105`, I have added a
specific `noqa` ignore for that.