Infractions: fix UniqueTogetherValidator incorrectly failing
The active infractions queryset only gets filtered by the fields
specified. This meant that if the same user and type had another
infraction instance which was active, the validator would fail. The
validator assumes failure if it sees any items still in the queryset
after filtering.

By including the active field in the validator, the queryset will be
filtered by the active field too. In the case described above, the
queryset would end up empty because a no infractions which are active
will ever match an active=False filter.