Remove default value from the active field of infractions
Due to the active field being specified in the UniqueTogetherValidator,
the field is implicitly required. Typically default values are excluded
from this restriction but in this case some infraction types must always
be False.

It's easier and makes more sense to require the active field explicitly
rather than to write logic in the serializer which is conditional on the
type of infractions.