Allow empty list for message embeds
By default, blank=False for ArrayFields but allow_empty=True for
ListField. Before DRF 3.10 there was a bug that ListField didn't respect
the default value of blank=False and thus created a ListField in the
serialiser with the default of allow_empty=True. We were relying on the
behaviour of that bug.

See encode/django-rest-framework#6597.