Handle unique validator in DRF, not Django.
I was handling this in a Django vanilla kind of way, which was causing
the constraint to return a 500 instead of a 400. This changes the
approach to use the DRF way, and makes it return 400.

It doesn't actually change the way anything behaves, other than
returning the right status code.