Remove delete permission for bot settings admin.
I'm unable to see any cases where this would be wanted, and instead accidental deletion would result in the system possibly breaking, as we are unable to add the setting again to replace it if it got removed.

The name has also set to read only in item view, to prevent renames, effectively doing the same thing as deleting it.