Make last_updated field automatically update.
Previously, we were operating under a bad assumption that we would be
updating the last_updated field in the RepositoryMetadata objects
whenever we updated the objects with new data from the GitHub API.

Upon closer inspection, this is not at all what we're doing, and some of
those repository objects had not been updated in over a year!

This introduces `auto_now` to the field, which will ensure that it is
automatically updated whenever the object is updated.