Fix misleading coverage report.
Due to an optimization in CPython that is amended in 3.10, coverage.py
is sometimes unable to determine the coverage of continue statements in
branches. See: https://github.com/nedbat/coveragepy/issues/198

Adding a no-op like a print or an empty statement would solve the
coverage issue, but I've opted to just ignore the line. This should be
tested and the line removed when the site is updated to Python 3.10.