Fix `content` app tests not running on macOS.
macOS uses `/var/...` as its temp directory, causing issues with the hardcoded
usage of `/tmp` as the temporary directory. As `/tmp` is also used in some
tests (for now), population of the directory is done regardless.

I have also attempted completely removing the `/tmp` path in tests in favour of
the platform-specific temporary directory, however, with the current
implementation if directory traversal for wiki content, this is currently more
complicated to implement (in the tests) than this solution.