Fix `content` app tests not running on macOS (#519)
macOS uses `/var/...` as its temp directory, causing issues with the
hardcoded usage of `/tmp` as the temporary directory. Therefore,
relying on tmp is not  portable. 

Populating the true temporary directory is redundant and may cause more
problems because of nested directories. Move the fake content under a
subdirectory to avoid this issue.

Co-authored-by: MarkKoz <KozlovMark@gmail.com>