Create db then switch conn, rather than use dblink
The previous implementation would succesfully create the metricity db, but the tables and rows would be inserted into pysite.

This is because dblink doesn't switch contexts within the script, as psql doesn't support this.

If we were to stick with 1 sql file, we would need to wrap every single statement into a dblink_exec() call to force psql to exec on the other db.