Restructure bot contributing guide
Restructures the bot guide to focus on getting users "through the door" first, and slowly adding additional information and options.
The guide is built in terms of showing ways to run the bot and how to achieve it, rather than listing all of the information for each configuration and environment variable right away.
That said, the attempt was to still present the reader with the full information they need, and to still be able to use as a quick reference for contributors who are already familiar with the project.

As a style choice, this guide incorporates collapsible elements, which required some CSS and JS additions. This could be resolved with the bulma-collapsible extension, but it doesn't seem to function well inside django templates.

The guide includes an optional config.yml to use with the test server template. This is a temporary measure until we get the bootstrapping application going.

This guide additionally splits the bot account creation and obtaining Discord ID's guides and they didn't seem related. The original file is kept for now as it is used by other guides at the moment.