Fix typos across codebase
./pydis_site/apps/resources/resources/tools/ides/thonny.yaml:1: specifically
./pydis_site/apps/content/resources/guides/pydis-guides/helping-others.md:115: considered
./pydis_site/apps/content/resources/guides/pydis-guides/contributing/issues.md:59 labels
./pydis_site/apps/content/resources/guides/pydis-guides/contributing/sir-lancebot.md:99: recommend
./pydis_site/apps/content/resources/guides/pydis-guides/contributing/site.md:111: particularly
./pydis_site/apps/content/resources/guides/pydis-guides/contributing/sir-lancebot/env-var-reference.md:29: Integer
./pydis_site/apps/content/resources/guides/pydis-guides/contributing/sir-lancebot/env-var-reference.md:67: calculating
./pydis_site/apps/api/tests/test_off_topic_channel_names.py:157: response

I didn't touch the code jam and game jam typos because I'm not sure if
they should be preserved as is.

There were a few 'seperated' typos which I didn't change because I
*think* it's just another way of spelling it?

In the offensive words test there was a keyword argument named `fied`
which I didn't touch because I wasn't sure where that was from.