Only calc activity blocks when <1k messages
We only truly care about how many activity blocks a user has when they have a small number of messages, as we only use this for the voice gate.

If a user has more than say 1k messages, then we really don't need to calculate their activity blocks, as it's quite an expensive query with many messages.