Merged infraction and notification settings in JSON
The settings for infracting and notifying the user were merged under one field, which is renamed to "infraction_and_notification".
The only place which sends a message in the server by default is the antimalware, the rest try to DM the user first, and antimalware can do the same.

This avoids complications which may result from the filtering cog trying to send two messages: one for the defined server message, and another for a failed DM.