Fix resource redirects for deploy previews.
This could probably be improved by making it less hardcoded, but since
there's a bunch of special cases, and since it only affects deploy
previews, I'm choosing to solve this with a simple solution.

I don't think realistically that we're gonna be adding a ton of resource
types down the line anyway.