Rewrite to emphasize GPU computation
Previously, it was unclear why this was a data science resource. This version states that Colab offers GPU computation and that it can be used for deep learning, which is the reason. Also fixes abrupt change in sentence structure ("a ... Jupyter Notebook which supports e.g. !apt-get"; "e.g." unexpected).