Rewrite Google Colab resource to emphasize GPU computation
Previously, it was unclear why this was a data science resource. This version states that Colab offers GPU computation and that it can be used for deep learning, which is the reason. Also fixes abrupt change in sentence structure ("a ... Jupyter Notebook which supports e.g. !apt-get"; "e.g." unexpected).

Also, expressions in backticks (apt-get) won't be rendered as code, though since Colab is mostly a beginner sandbox, I don't think the word "Debian" will often be known.