Remove Message Model Test
The message model was tested by instantiating and confirming it has a
string representation, but instantiating abstract models is undefined
behavior, and can break with future versions of django.

The behavior of the test is redundant anyway, since an abstract model
wouldn't exist in isolation, and the desired behavior is confirmed by
inheritors.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>