Bump Django To 3.2
Bumps the current django version to 3.2 For the most part, there are no
major changes here. The app configuration names were incorrect, and the
new version no longer ignores that, so those were updated. The new
version also requires explicitly defining the field type for primary
keys if they are not defined on the model.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>