Explicitly Specify Infraction Time In Tests
The infraction tests checked that the route returned infractions in the
correct order, which is based on insertion time. This can be fragile
however, since the insertion time can be very close (or identical)
during the tests. That became especially more likely with PR #741
(commit 149e67b4) which improved database access speed.

This is fixed by explicitly specifying the insertion time, and spacing
them out properly.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>