Fix Django Deprecation Warnings
Removes a few features which were deprecated in django 4.0. Running with
warnings enabled shows no other errors.

USE_L10N was deprecated, and is now enabled by default. In future
versions of django, it'll be impossible to turn localization off.

Explicitly defining the custom_app_config for the API app is no longer
necessary as django can pick it up on its own.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>