Bump Developer Dependencies
Most of these bumps are just locking to exact versions since the
latest version for most packages were already within the allowed
range.

flake8-bandit had to be disabled temporarily as it is currently broken
for the latest version of flake8. The PR fixing the issue is linked,
and it can be re-enabled when a new version is available.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>