Switch To Coveralls Action
The coveralls package has consistently caused issues for us, so it made
sense to switch to something that's a little more decoupled, especially
when it has official support.

Switching is simple, since coverage.py can natively generate the format
used by the coveralls action, and everything else is plug and play.

Signed-off-by: Hassan Abouelela <hassan@hassanamr.com>