Drop dependency to pyfakefs
Create a temporary directory to manage our resource tests instead of
reyling on pyfakefs to mock it away for us. This also makes the code
more portable: all we need now is a way to create a temporary directory.
`pathlib` mostly abstracts away the other parts for us. Since we're
well-behaved, we clean up the temporary directory at the end of the
Python interpreter's life using `atexit` and `shutil.rmtree`.

This PR was written and tested with Python 3.9 which required some hacks
in `pyproject.toml` to make it work, it may require re-locking if CI
throws up.

Closes #679.