Remove the redirect FilterList on migration
This redirect filter list does not have any filters, nor does it have a bot implementation.
It is something that we were working on, but has been stalled for a while.

The presence of this filter list also causes the bot to warn on startup due to it finding a filter list with no implementation.
This commit removes the FilterList, which can be added back if/when we support this filter type in bot.