Timeline: Migrate to Markdown source files - initial implementation
- The timeline app is introduced, moved from the home app.
- Add a basic README.md to illustrate the overall code breakdown of the
  app as it is in the current stage.
- Two entries are added for now.
- Add ability to link headers using the slug portion of the filename.
- The way markdown files are fetched is similar to that of the resources
  app - using the AppConfig ready() function, all resources are ensured
  to be only fetched once when the app is first laoded.

I debated whether to introduce the new functionality in the home app
instead, without creating a new app. Eventually I decided extracting it
to a standalone app now is better as it allows easier extension of
functionality in the future. The home app can remain as it is to only
server the `/` homepage.

The timeline CSS is kept the same, as with the structure of the timeline
HTML template.

Once the CSS rewrite PR is merged, it's relatively easy to fix conflicts
here (again, since timeline is now its own app, with the CSS file and
HTML template moved - extra advantage).

References to `home:timeline` are updated to use `timeline:index`
throughout the codebase, as far as my ripgrep search could go.

The format of the markdown + YAML entries are still up for debate, so
I've only added the first two entries for now. They can be completely
overwritten in the future once the formats are decided by using a script
to convert all the data from my JSON file into individual markdown
files:

    http://0x0.st/Xr78.txt

This link should last for at least a few months. (Originally saved on
https://paste.pythondiscord.com/KPJA, but it expires on 12th April 2024.)